+++ packages/quantic/force-app/main/default/lwc/quanticSearchBox/quanticSearchBox.js
@@ -7,9 +7,16 @@ import {
 
 import search from '@salesforce/label/c.quantic_Search';
 
-const ENTER = 13;
-const ARROWUP = 38;
-const ARROWDOWN = 40;
+const KEYS = {
+  ENTER: 'Enter',
+  ARROWUP: 'ArrowUp',
+  ARROWDOWN: 'ArrowDown',
+};
+
+const CLASS_WITH_SUBMIT =
+  'slds-combobox__form-element slds-input-has-icon slds-input-has-icon_right slds-input-has-fixed-addon';
+const CLASS_WITHOUT_SUBMIT =
+  'slds-combobox__form-element slds-input-has-icon slds-input-has-icon_left-right';
 
 export default class QuanticSearchBox extends LightningElement {
   /** @type {import("coveo").SearchBoxState} */
@@ -19,10 +26,12 @@ export default class QuanticSearchBox extends LightningElement {
     suggestions: [],
     value: '',
   };
+
   /** @type {any} */
   get suggestions() {
-    return this.state.suggestions.map((s, index) => ({
+    return this.searchBox.state.suggestions.map((s, index) => ({
       key: index,
+      rawValue: s.rawValue,
       value: s.highlightedValue,
     }));
   }
@@ -40,33 +49,11 @@ export default class QuanticSearchBox extends LightningElement {
   /** @type {number} */
   @api numberOfSuggestions = 5;
 
-  /** @type {string} */
-  searchBoxContainerClass;
   /** @type {import("coveo").SearchBox} */
   searchBox;
   /** @type {import("coveo").Unsubscribe} */
   unsubscribe;
 
-  /** @type {number} */
-  selectionIndex = -1;
-  /** @type {HTMLInputElement} */
-  input;
-  /** @type {HTMLElement} */
-  combobox;
-  /** @type {HTMLButtonElement} */
-  clearButton;
-  /** @type {() => void} */
-  resetSelectionIndex = () => {
-    this.selectionIndex = -1;
-  };
-  /** @type {() => boolean} */
-  areSuggestionsShown = () =>
-    this.template
-      .querySelector('.slds-combobox')
-      .classList.contains('slds-is-open');
-  /** @type {() => NodeListOf<HTMLElement>} */
-  getSuggestionElements = () =>
-    this.template.querySelectorAll('.slds-listbox__option');
 
   /**
    * @param {import("coveo").SearchEngine} engine
@@ -93,16 +80,6 @@ export default class QuanticSearchBox extends LightningElement {
 
   renderedCallback() {
     initializeWithHeadless(this, this.engineId, this.initialize.bind(this));
-    if (!this.input) {
-      this.input = this.template.querySelector('input');
-    }
-    if (!this.combobox) {
-      this.combobox = this.template.querySelector('.slds-combobox');
-    }
-    if (!this.clearButton) {
-      this.clearButton = this.template.querySelector('.slds-button__icon');
-    }
-    this.setSearchBoxContainerClass();
   }
 
   disconnectedCallback() {
@@ -118,55 +95,63 @@ export default class QuanticSearchBox extends LightningElement {
     this.state = this.searchBox.state;
   }
 
-  setSearchBoxContainerClass() {
-    const withoutSubmitClass =
-      'slds-combobox__form-element slds-input-has-icon slds-input-has-icon_left-right';
-    const withSubmitClass =
-      'slds-combobox__form-element slds-input-has-icon slds-input-has-icon_right slds-input-has-fixed-addon';
+  /**
+   * @returns {import('c/quanticSearchBoxSuggestionsList').default}
+   */
+  get suggestionList() {
+    // @ts-ignore
+    return this.template.querySelector('c-quantic-search-box-suggestions-list');
+  }
+
+  /**
+   * @returns {HTMLInputElement}
+   */
+  get input() {
+    return this.template.querySelector('input');
+  }
+
+  /**
+   * @returns {HTMLElement}
+   */
+  get combobox() {
+    return this.template.querySelector('.slds-combobox');
+  }
+
+  get searchBoxContainerClass() {
     if (this.withoutSubmitButton) {
-      this.searchBoxContainerClass = withoutSubmitClass;
-      this.input.setAttribute('aria-labelledby', 'fixed-text-label');
-    } else {
-      this.searchBoxContainerClass = withSubmitClass;
-      this.input.setAttribute(
-        'aria-labelledby',
-        'fixed-text-label fixed-text-addon-post'
-      );
+      this.input?.setAttribute('aria-labelledby', 'fixed-text-label');
+      return CLASS_WITH_SUBMIT;
     }
+    this.input?.setAttribute(
+      'aria-labelledby',
+      'fixed-text-label fixed-text-addon-post'
+    );
+    return CLASS_WITHOUT_SUBMIT;
+  }
+
+  get suggestionsOpen() {
+    return this.combobox.classList.contains('slds-is-open');
+  }
+
+  get isQueryEmpty() {
+    return !this.input?.value?.length;
   }
 
   showSuggestions() {
-    this.combobox.classList.add('slds-is-open');
-    this.combobox.setAttribute('aria-expanded', 'true');
+    this.searchBox.showSuggestions();
+    this.combobox?.classList.add('slds-is-open');
+    this.combobox?.setAttribute('aria-expanded', 'true');
   }
 
   hideSuggestions() {
-    this.combobox.classList.remove('slds-is-open');
-    this.combobox.setAttribute('aria-expanded', 'false');
-    this.resetHighlighted();
-    this.resetSelectionIndex();
+    this.combobox?.classList.remove('slds-is-open');
+    this.combobox?.setAttribute('aria-expanded', 'false');
+    this.suggestionList?.resetSelection();
   }
 
-  setHighlighted() {
-    const suggestions = this.getSuggestionElements();
-    if (!suggestions.length) {
-      return;
-    }
-    const suggestion = suggestions[this.selectionIndex];
-
-    this.resetHighlighted();
-    suggestion.setAttribute('aria-selected', 'true');
-    suggestion.classList.add('slds-has-focus');
-    this.input.value = suggestion.innerText;
-  }
-
-  resetHighlighted() {
-    const options = this.getSuggestionElements();
-
-    options.forEach((element) => {
-      element.setAttribute('aria-selected', 'false');
-      element.classList.remove('slds-has-focus');
-    });
+  handleHighlightChange(event) {
+    const suggestion = event.detail;
+    this.input.value = suggestion.rawValue;
   }
 
   /**
@@ -174,35 +159,24 @@ export default class QuanticSearchBox extends LightningElement {
    */
   updateSearchboxText(textValue) {
     this.input.value = textValue;
-    this.searchBox.updateText(textValue);
   }
 
   handleEnter() {
-    if (this.selectionIndex >= 0) {
+    const selectedSuggestion = this.suggestionList?.getCurrentSelectedValue();
+    if (this.suggestionsOpen && selectedSuggestion) {
+      this.searchBox.selectSuggestion(selectedSuggestion.rawValue);
+      this.input.blur();
+    } else {
+      this.searchBox.submit();
+      this.input.blur();
+    }
+  }
+
+  onSubmit(event) {
+    event.stopPropagation();
+    if(this.searchBox.state.value !== this.input.value) {
       this.searchBox.updateText(this.input.value);
     }
-    this.searchBox.submit();
-    this.input.blur();
-  }
-
-  handleArrowUp() {
-    this.selectionIndex--;
-    if (this.selectionIndex < 0) {
-      this.selectionIndex = this.suggestions.length - 1;
-    }
-    this.setHighlighted();
-  }
-
-  handleArrowDown() {
-    this.selectionIndex++;
-    if (this.selectionIndex > this.suggestions.length - 1) {
-      this.selectionIndex = 0;
-    }
-    this.setHighlighted();
-  }
-
-  onSubmit() {
-    this.searchBox.updateText(this.input.value);
     this.searchBox.submit();
     this.input.blur();
   }
@@ -211,28 +185,23 @@ export default class QuanticSearchBox extends LightningElement {
    * @param {KeyboardEvent & {target: {value : string}}} event
    */
   onKeyup(event) {
-    if (this.input.value === '') {
-      this.clearButton.classList.remove('slds-visible');
-      this.clearButton.classList.add('slds-hidden');
-    } else {
-      this.clearButton.classList.remove('slds-hidden');
-      this.clearButton.classList.add('slds-visible');
-    }
-    if (event.which === ENTER) {
-      this.handleEnter();
-    } else if (this.areSuggestionsShown() && event.which === ARROWUP) {
-      this.handleArrowUp();
-    } else if (this.areSuggestionsShown() && event.which === ARROWDOWN) {
-      this.handleArrowDown();
-    } else {
-      this.resetSelectionIndex();
-      this.resetHighlighted();
-      this.searchBox.updateText(event.target.value);
+    switch (event.key) {
+      case KEYS.ENTER:
+        this.handleEnter();
+        break;
+      case KEYS.ARROWUP:
+        this.suggestionList?.selectionUp();
+        break;
+      case KEYS.ARROWDOWN:
+        this.suggestionList?.selectionDown();
+        break;
+      default:
+        this.suggestionList?.resetSelection();
+        this.searchBox.updateText(event.target.value);
     }
   }
 
   onFocus() {
-    this.searchBox.showSuggestions();
     this.showSuggestions();
   }
 
@@ -242,20 +211,13 @@ export default class QuanticSearchBox extends LightningElement {
 
   clearInput() {
     this.input.value = '';
-    this.clearButton.classList.remove('slds-visible');
-    this.clearButton.classList.add('slds-hidden');
     this.searchBox.updateText(this.input.value);
     this.input.focus();
   }
 
-  preventDefault(event) {
-    event.preventDefault();
-  }
-
   handleSuggestionSelection(event) {
-    const textValue = event.target.innerText;
-    this.updateSearchboxText(textValue);
-    this.searchBox.submit();
+    const textValue = event.detail;
+    this.searchBox.selectSuggestion(textValue);
     this.input.blur();
   }
 }